refactor: delete Base ViewController
