+++ GithubClient/Controllers/MainTabBarViewController.swift
@@ -36,11 +36,11 @@ class MainTabBarViewController: UITabBarController {
         
         guard let githubModel = self.githubModel else { return }
         
-        let firstViewController = RepoViewController(tableView: RepoTableView(), githubModel: githubModel)
+        let firstViewController = RepoViewController(tableView: RepoTableView(), githubModel: githubModel, loadingview: LoadingView())
         firstViewController.tabBarItem = UITabBarItem(title: Const.NavigationTitle.repoPage, image: UIImage(systemName: "list.bullet"), selectedImage: UIImage(systemName: "list.bullet"))
         viewcontrollers.append(firstViewController)
         
-        let secondViewController = ProfileViewController(githubModel: githubModel)
+        let secondViewController = ProfileViewController(githubModel: githubModel, loadingView: LoadingView())
         secondViewController.tabBarItem = UITabBarItem(title: Const.NavigationTitle.profilePage, image: UIImage(systemName: "person"), selectedImage: UIImage(systemName: "person"))
         viewcontrollers.append(secondViewController)
