+++ GithubClient/Controllers/ProfileViewController.swift
@@ -6,25 +6,73 @@
 //
 
 import UIKit
+import RxSwift
 
-class ProfileViewController: BaseViewController {
+class ProfileViewController: UIViewController {
     
+    // MARK: field variable
     var userData: UserEntity? = nil
     
+    private weak var delegate: BaseViewDelegate?
     private var githubModel: GithubModel?
+    private var subscription: Disposable?
+    private var loadingView: LoadingView?
     
-    init(githubModel: GithubModel) {
+    init(githubModel: GithubModel, loadingView: LoadingView) {
         self.githubModel = githubModel
-        super.init(loadingView: LoadingView(), requestUrl: "https://api.github.com/users/fucchi-senpai")
+        self.loadingView = loadingView
+        super.init(nibName: nil, bundle: nil)
     }
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
     
+    // MARK: Life cycle function
     override func viewDidLoad() {
-        self.delegate = self
         super.viewDidLoad()
+        self.delegate = self
+        self.beforeLoad()
+        self.delegate?.load(url: "https://api.github.com/users/fucchi-senpai") { data in
+            self.delegate?.setUp(data: data)
+            self.postLoad()
+        }
+    }
+    
+    open override func viewDidDisappear(_ animated: Bool) {
+        self.subscription?.dispose()
+    }
+    
+    // MARK: Private function
+    /// API読み込み前処理
+    private func beforeLoad() {
+        print(#function)
+        DispatchQueue.main.async {
+            self.view.backgroundColor = .systemBackground
+            self.initLoadingView()
+            self.loadingView?.start()
+        }
+    }
+    
+    /// API読み込み後処理
+    private func postLoad() {
+        print(#function)
+        DispatchQueue.main.async {
+            self.delegate?.initViews()
+            self.loadingView?.stop()
+        }
+    }
+    
+    private func initLoadingView() {
+        guard let loadingView = self.loadingView else { return }
+        self.view.addSubview(loadingView)
+        loadingView.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            loadingView.heightAnchor.constraint(equalTo: self.view.heightAnchor, multiplier: 0.2),
+            loadingView.widthAnchor.constraint(equalTo: self.view.widthAnchor, multiplier: 0.2),
+            loadingView.centerYAnchor.constraint(equalTo: self.view.centerYAnchor),
+            loadingView.centerXAnchor.constraint(equalTo: self.view.centerXAnchor)
+        ])
     }
     
     private func initNavigationView() {
